Split dashboard proof view modules
